Move header redaction to its own file
